backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
dj-database-url==2.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
channels==4.1.0
daphne==4.1.2